added PV generating templates

+++ src/main/epics/tspd-adqApp/src/ADQAIChannel.h
@@ -12,18 +12,15 @@ public:
 
     void setState(nds::state_t newState);
 
-    //// urojec L3: maybe make these methods a bit more descriptive. It is
-    ////            a little bit hard to know what they do just from looking at
-    ////            the name
     void readData(short* rawData, std::int32_t sampleCnt);
-    void getDataPV(timespec* pTimestamp, std::vector<double>* pValue);
+    void getDataPV(timespec* pTimestamp, std::vector<int8_t>* pValue);
 
     void commitChanges(bool calledFromDaqThread = false);
 
 private:
     nds::Node m_node;
     nds::StateMachine m_stateMachine;
-    std::vector<double> m_data;
+    std::vector<int8_t> m_data;
 
     bool m_firstReadout;
 
@@ -34,8 +31,7 @@ private:
     void recover();
     bool allowChange(const nds::state_t, const nds::state_t, const nds::state_t);
 
-    //nds::PVDelegateIn<std::vector<uint8_t>> m_dataPV;
-    nds::PVDelegateIn<std::vector<double>> m_dataPV;
+    nds::PVDelegateIn<std::vector<int8_t>> m_dataPV;
 };
 
 #endif /* ADQAICHANNEL_H; */
